+++ src/main.ts
@@ -1,29 +1,25 @@
 import ModalController from "./modal/ModalController";
 import BodyClick from "./dom/BodyClick";
 import BodyScroll from "./dom/BodyScroll";
-import { ModalCallback, ModalConfig, ModalEvent } from "./types";
+import { ModalCallback, ModalConfig, ModalEventMap } from "./interfaces";
 
 const bodyClick = new BodyClick();
 const bodyScroll = new BodyScroll();
 
-document.addEventListener("opening:modal", (event: ModalEvent) => {
+document.addEventListener("opening:modal", (event) => {
   // Removes overflow hidden from the body
-  console.log("opening:modal");
   bodyScroll.lock(event.detail);
 });
-
-document.addEventListener("closed:modal", (event: ModalEvent) => {
-  console.log("closed:modal");
+document.addEventListener("closed:modal", (event) => {
   // Sets overflow hidden to the body
   bodyScroll.unlock(event.detail);
 });
-
 document.addEventListener("DOMContentLoaded", () => {
   // BodyClick class uses single event listener, no meter how many modals we have
   document.body.addEventListener("click", bodyClick.eventHandler);
 });
 
-class Modal extends ModalController {
+class OpenModalJs extends ModalController {
   constructor(modalId: string, config?: ModalConfig, callback?: ModalCallback) {
     super(modalId, config, callback);
 
@@ -32,5 +28,19 @@ class Modal extends ModalController {
   }
 }
 
-// Since this is vanilla JS, we are adding the Modal class globally
-window.Modal = Modal;
+declare global {
+  // eslint-disable-next-line no-unused-vars
+  interface Window {
+    OpenModalJs: typeof OpenModalJs;
+  }
+  interface Document {
+    addEventListener<K extends keyof ModalEventMap>(
+      type: K,
+      listener: (this: Document, ev: ModalEventMap[K]) => void
+    ): void;
+  }
+}
+
+export { ModalController, BodyClick, BodyScroll };
+export default OpenModalJs;
+window.OpenModalJs = OpenModalJs;