+++ types/dom/BodyClick.d.ts
@@ -1,0 +1,9 @@
+import { ModalInterface } from "../interfaces";
+export default class BodyClick {
+    #private;
+    private readonly modals;
+    readonly eventHandler: OmitThisParameter<(event: MouseEvent) => void>;
+    constructor();
+    addModalToHandler(modal: ModalInterface): void;
+}
+//# sourceMappingURL=BodyClick.d.ts.map