+++ types/dom/BodyScroll.d.ts
@@ -1,0 +1,9 @@
+import { ModalInterface } from "../interfaces";
+export default class BodyScroll {
+    #private;
+    constructor();
+    get locked(): boolean;
+    lock(modal: ModalInterface): void;
+    unlock(modal: ModalInterface): void;
+}
+//# sourceMappingURL=BodyScroll.d.ts.map